changed width and height to be uint32_t instead of uint64_t

+++ kernel/framebuffer.cpp
@@ -3,10 +3,10 @@
 #include <stddef.h>
 
 uint32_t* fb_addr;
-uint64_t fb_width;
-uint64_t fb_height;
+uint32_t fb_width;
+uint32_t fb_height;
 
-void fb_init(uint32_t* addr, uint64_t width, uint64_t height) {
+void fb_init(uint32_t* addr, uint32_t width, uint32_t height) {
     fb_addr = addr;
     fb_width = width;
     fb_height = height;
@@ -22,12 +22,12 @@ void fb_clear() {
     }
 }
 
-void draw_pixel(const uint64_t x, const uint64_t y, const uint32_t color) {
+void draw_pixel(const uint32_t x, const uint32_t y, const uint32_t color) {
     const size_t pos = y * fb_width + x;
     fb_addr[pos] = color;
 }
 
-void draw_bitmap(const uint32_t* bitmap, const uint64_t x_pos, const uint64_t y_pos, const uint64_t width, const uint64_t height) {
+void draw_bitmap(const uint32_t* bitmap, const uint32_t x_pos, const uint32_t y_pos, const uint32_t width, const uint32_t height) {
     for (size_t y = 0; y < height; y++) {
         for (size_t x = 0; x < width; x++) {
             const size_t i = y * width + x;
